requirements.txt
fastapi
uvicorn[standard]
aiofiles
python-jose[cryptography]
passlib==1.7.4
bcrypt==4.3.0
python-multipart